Refactor Cypress tests to use new metadata function API
