+++ apps/ffdweb-site/cypress/e2e/about.cy.ts
@@ -1,14 +1,16 @@
 import { PATHS } from '@/constants/paths'
+import { BASE_URL } from '@/constants/siteMetadata'
 
 import { ABOUT_SEO } from '@/about/constants/seo'
-import { tests } from '@/support/index'
+import { tests } from '@/cypress/support'
 
 describe('About Page', () => {
   it(tests.metadata.prompt, () => {
     tests.metadata.fn({
       path: PATHS.ABOUT.path,
-      title: { absolute: ABOUT_SEO.title },
+      title: ABOUT_SEO.title,
       description: ABOUT_SEO.description,
+      baseUrl: BASE_URL,
     })
   })
