--- apps/ffdweb-site/cypress/e2e/digest_slug.ts
@@ -1,29 +1,0 @@
-import { PATHS } from '@/constants/paths'
-import { ORGANIZATION_NAME_SHORT } from '@/constants/siteMetadata'
-
-import type {
-  GenericEntryFrontmatter,
-  GetEntryOptions,
-} from '../tasks/getEntryFrontmatter'
-
-import { tests } from '@/support/index'
-
-const CONTENT_FOLDER = 'digest'
-
-describe('Digest Slug Page', () => {
-  it(tests.metadata.prompt, () => {
-    cy.task<string>('getRandomSlug', CONTENT_FOLDER).then((slug) => {
-      cy.task<GenericEntryFrontmatter>('getEntryFrontmatter', {
-        collection: CONTENT_FOLDER,
-        slug,
-      } satisfies GetEntryOptions).then(({ title, seo }) => {
-        const seoTitle = seo.title || title
-        tests.metadata.fn({
-          path: `${PATHS.DIGEST.path}/${slug}`,
-          title: { absolute: `${seoTitle} | ${ORGANIZATION_NAME_SHORT}` },
-          description: seo.description,
-        })
-      })
-    })
-  })
-})
