+++ apps/ffdweb-site/cypress/e2e/faqs.cy.ts
@@ -1,14 +1,16 @@
 import { PATHS } from '@/constants/paths'
+import { BASE_URL } from '@/constants/siteMetadata'
 
+import { tests } from '@/cypress/support'
 import { FAQS_SEO } from '@/faqs/constants/seo'
-import { tests } from '@/support/index'
 
 describe('FAQs Page', () => {
   it(tests.metadata.prompt, () => {
     tests.metadata.fn({
       path: PATHS.FAQS.path,
-      title: { absolute: FAQS_SEO.title },
+      title: FAQS_SEO.title,
       description: FAQS_SEO.description,
+      baseUrl: BASE_URL,
     })
   })
