+++ apps/ffdweb-site/cypress/e2e/learning_resources.cy.ts
@@ -1,14 +1,16 @@
 import { PATHS } from '@/constants/paths'
+import { BASE_URL } from '@/constants/siteMetadata'
 
+import { tests } from '@/cypress/support'
 import { LEARNING_RESOURCES_SEO } from '@/learning-resources/constants/seo'
-import { tests } from '@/support/index'
 
 describe('Learning Resources Page', () => {
   it(tests.metadata.prompt, () => {
     tests.metadata.fn({
       path: PATHS.LEARNING_RESOURCES.path,
-      title: { absolute: LEARNING_RESOURCES_SEO.title },
+      title: LEARNING_RESOURCES_SEO.title,
       description: LEARNING_RESOURCES_SEO.description,
+      baseUrl: BASE_URL,
     })
   })
