+++ apps/ffdweb-site/cypress/e2e/projects_index.cy.ts
@@ -1,14 +1,16 @@
 import { PATHS } from '@/constants/paths'
+import { BASE_URL } from '@/constants/siteMetadata'
 
+import { tests } from '@/cypress/support'
 import { PROJECTS_SEO } from '@/projects/constants/seo'
-import { tests } from '@/support/index'
 
 describe('Projects Index Page', () => {
   it(tests.metadata.prompt, () => {
     tests.metadata.fn({
       path: PATHS.PROJECTS.path,
-      title: { absolute: PROJECTS_SEO.title },
+      title: PROJECTS_SEO.title,
       description: PROJECTS_SEO.description,
+      baseUrl: BASE_URL,
     })
   })
