+++ apps/ffdweb-site/cypress/e2e/projects_slug.cy.ts
@@ -1,12 +1,12 @@
 import { PATHS } from '@/constants/paths'
-import { ORGANIZATION_NAME_SHORT } from '@/constants/siteMetadata'
+import { BASE_URL } from '@/constants/siteMetadata'
 
+import { tests } from '@/cypress/support'
 import type {
   GenericEntryFrontmatter,
   GetEntryOptions,
-} from '../tasks/getEntryFrontmatter'
-
-import { tests } from '@/support/index'
+} from '@/cypress/tasks/getEntryFrontmatter'
+import { getMetaTitleWithSuffix } from '@/cypress/utils/getMetaTitleWithSuffix'
 
 const CONTENT_FOLDER = 'projects'
 
@@ -21,8 +21,9 @@ describe('Projects Slug Page', () => {
 
         tests.metadata.fn({
           path: `${PATHS.PROJECTS.path}/${slug}`,
-          title: { absolute: `${seoTitle} | ${ORGANIZATION_NAME_SHORT}` },
+          title: getMetaTitleWithSuffix(seoTitle),
           description: seo.description,
+          baseUrl: BASE_URL,
         })
       })
     })
