+++ apps/ffdweb-site/cypress/utils/getMetaTitleWithSuffix.ts
@@ -1,0 +1,5 @@
+import { ORGANIZATION_NAME_SHORT } from '@/constants/siteMetadata'
+
+export function getMetaTitleWithSuffix(title: string) {
+  return `${title} | ${ORGANIZATION_NAME_SHORT}`
+}
